# database/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package db

import (
	"time"

	"github.com/google/uuid"
	"github.com/jackc/pgx/v5/pgtype"
)

type Branch struct {
	ID           uuid.UUID          `json:"id"`
	Name         string             `json:"name"`
	CompanyID    uuid.UUID          `json:"company_id"`
	Latitude     pgtype.Float8      `json:"latitude"`
	Longitude    pgtype.Float8      `json:"longitude"`
	LocationName pgtype.Text        `json:"location_name"`
	CreatedAt    pgtype.Timestamptz `json:"created_at"`
	UpdatedAt    pgtype.Timestamptz `json:"updated_at"`
}

type Company struct {
	ID           uuid.UUID          `json:"id"`
	Name         string             `json:"name"`
	Latitude     pgtype.Float8      `json:"latitude"`
	Longitude    pgtype.Float8      `json:"longitude"`
	LocationName pgtype.Text        `json:"location_name"`
	CreatedAt    pgtype.Timestamptz `json:"created_at"`
	UpdatedAt    pgtype.Timestamptz `json:"updated_at"`
}

type Department struct {
	ID        uuid.UUID          `json:"id"`
	Name      string             `json:"name"`
	CompanyID uuid.UUID          `json:"company_id"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
	UpdatedAt pgtype.Timestamptz `json:"updated_at"`
}

type Employee struct {
	ID           uuid.UUID          `json:"id"`
	UserID       uuid.UUID          `json:"user_id"`
	Salary       float64            `json:"salary"`
	Status       string             `json:"status"`
	CompanyID    uuid.UUID          `json:"company_id"`
	BranchID     pgtype.UUID        `json:"branch_id"`
	DepartmentID pgtype.UUID        `json:"department_id"`
	SiteID       pgtype.UUID        `json:"site_id"`
	CreatedAt    pgtype.Timestamptz `json:"created_at"`
	UpdatedAt    pgtype.Timestamptz `json:"updated_at"`
}

type Site struct {
	ID           uuid.UUID          `json:"id"`
	Name         string             `json:"name"`
	BranchID     uuid.UUID          `json:"branch_id"`
	Latitude     pgtype.Float8      `json:"latitude"`
	Longitude    pgtype.Float8      `json:"longitude"`
	LocationName pgtype.Text        `json:"location_name"`
	CreatedAt    pgtype.Timestamptz `json:"created_at"`
	UpdatedAt    pgtype.Timestamptz `json:"updated_at"`
}

type User struct {
	ID                uuid.UUID          `json:"id"`
	FirstName         string             `json:"first_name"`
	MiddleName        string             `json:"middle_name"`
	LastName          string             `json:"last_name"`
	Username          string             `json:"username"`
	Password          string             `json:"password"`
	Phone             string             `json:"phone"`
	Email             string             `json:"email"`
	ProfilePictureUrl pgtype.Text        `json:"profile_picture_url"`
	CreatedAt         time.Time          `json:"created_at"`
	UpdatedAt         time.Time          `json:"updated_at"`
	DeletedAt         pgtype.Timestamptz `json:"deleted_at"`
}